dcp-845 move spreadsheet package to terra
