+++ exporter/terra/spreadsheet/config.py
@@ -4,7 +4,6 @@ from threading import Thread
 from hca_ingest.api.ingestapi import IngestApi
 
 from exporter.ingest.service import IngestService
-from exporter.ingest.spreadsheet.handler import SpreadsheetHandler
 from exporter.queue.config import QueueConfig, AmqpConnConfig
 from exporter.queue.connector import QueueConnector
 from exporter.queue.listener import QueueListener
@@ -13,7 +12,7 @@ from exporter.terra.experiment.client import TerraStorageClient
 from exporter.terra.experiment.config import TerraConfig
 from exporter.terra.gcs.config import GcpConfig
 from exporter.terra.gcs.storage import GcsStorage
-from exporter.terra.submission.client import TerraTransferClient
+from .handler import SpreadsheetHandler
 
 EXCHANGE = 'ingest.exporter.exchange'
 SPREADSHEET_QUEUE_CONFIG = QueueConfig(
